fixed bug 'getStaticResource is not a function'
