+++ .vscode/launch.json
@@ -9,8 +9,8 @@
             "type": "node",
             "request": "launch",
             "name": "Launch Program",
-            "program": "${workspaceFolder}/app/index.js",
-            "runtimeExecutable": "/home/lowe/Git/nvm/versions/node/v8.11.1/bin/node"
+            "program": "${workspaceFolder}/test/app/index.js",
+            "runtimeExecutable": "E:\\git\\nvmw\\v8.11.1\\node"
         }
     ]
 }
