--- config/proxy_table.js
@@ -1,8 +1,0 @@
-'use strict';
-
-module.exports = {
-    'api': {
-        host: 'localhost',
-        port: 3000
-    }
-};
