+++ test/views/index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Document</title>
+    <script src="https://cdn.bootcss.com/jquery/3.3.1/jquery.min.js"></script>
+    <script src="/views/test.js"></script>
+</head>
+<body>
+    <button onclick="send()">send</button>
+    <script>
+        function send() {
+            $.post('/api/banner',{a: 1}).then(()=>{
+            
+            })
+        }
+    </script>
+</body>
+</html>
